Cleaned usage of PBO in RenderSystem::CaptureRenderToBuffer. This is done in order to fix issue 4522. PBO was originally added for lightgem capturing for issue 4395.
Now CaptureRenderToBuffer accepts a flag (false by default), telling it to use PBO.
Also, PBO is allocated with proper size (not hardcoded magic constant).
Each time PBO is used, we check that its size is exactly the same as before.

Note: if some other part of code needs PBO,
then we can store all used PBOs in a map and specify which PBO to use by the usePbo argument.
Also, we can allow resizing PBO-s.
All of this is not used now, and perhaps won't be used in future.