Fixed some long-s in idPVS.
This may fix incorrect data in PVS on Linux 64-bit.
The function idPVS::Init assumes that long is 32-bit: it only rounds number of words in bitset up to 32-bit.
As a result, if total number of areas/portals is (64*q+r) for 1<=r<=32, then the last r bits are not processed in all bitset operations.