Replaced MD4 hash implementation.
New implementation was written by Alexander Peslyak:
  https://openwall.info/wiki/people/solar/software/public-domain-source-code/md4
It is in public domain, so we can do What The Fuck We Want with it.
Unlike the original version, which came directly from RSA and required explicit notice in any distribution.

MD4 is used:
  1) to calculate hash of all loaded pk4 (once during start)
  2) to calculate hash of game scripts (on every game save/load)
I have checked that the new code generates same result as the old one =)